+++ components/ui/Header.tsx
@@ -1,86 +1,33 @@
-import { ThemedView } from "@/components/ThemedView";
-import {
-  Modal,
-  SectionList,
-  StyleSheet,
-  Switch,
-  TouchableOpacity,
-  View,
-} from "react-native";
+import { StyleSheet, Switch, TouchableOpacity, View } from "react-native";
 import { useApp } from "@/context/AppContext";
 import { Text } from "@/components/ThemedText";
 import sizes from "@/config/sizes";
 import { MaterialCommunityIcons } from "@expo/vector-icons";
 import { useState } from "react";
-import RadioButton from "@/components/RadioButton";
 import { useThemeColor } from "@/hooks/useThemeColors";
-
-const data: { title: string; data: ("auto" | "dark" | "light")[] }[] = [
-  { title: "Appearance", data: ["auto", "dark", "light"] },
-];
 
 function Header() {
   const { isRounding, setIsRounding } = useApp();
-  const [modalVisible, setModalVisible] = useState<boolean>(false);
-
   const secColor = useThemeColor({}, "secondary");
   const accentColor = useThemeColor({}, "accent");
-  const whiteColor = useThemeColor({}, "white");
-  const textColor = useThemeColor({}, "text");
 
   return (
-    <>
-      <View style={styles.header}>
-        <View style={styles.roundContainer}>
-          <Switch
-            value={isRounding}
-            style={styles.switch}
-            trackColor={{
-              false: secColor,
-              true: accentColor,
-            }}
-            onValueChange={() => setIsRounding(!isRounding)}
-          />
-          <Text type="defaultSemiBold" style={[styles.roundText]}>
-            Round Up
-          </Text>
-        </View>
-        <TouchableOpacity onPress={() => setModalVisible(true)}>
-          <MaterialCommunityIcons
-            name="cog"
-            size={sizes.flg}
-            color={textColor}
-          />
-        </TouchableOpacity>
+    <View style={styles.header}>
+      <View style={styles.roundContainer}>
+        <Switch
+          value={isRounding}
+          style={styles.switch}
+          trackColor={{
+            false: secColor,
+            true: accentColor,
+          }}
+          onValueChange={() => setIsRounding(!isRounding)}
+        />
+        <Text type="defaultSemiBold" style={[styles.roundText]}>
+          Round Up
+        </Text>
       </View>
-      <Modal
-        visible={modalVisible}
-        animationType={"slide"}
-        transparent={true}
-        onRequestClose={() => {
-          setModalVisible(!modalVisible);
-        }}>
-        <ThemedView style={[{ flex: 1 }]}>
-          <TouchableOpacity
-            onPress={() => setModalVisible(false)}
-            style={{ alignSelf: "flex-end", margin: 15 }}>
-            <MaterialCommunityIcons
-              name="close"
-              size={sizes.fxl}
-              color={whiteColor}
-            />
-          </TouchableOpacity>
-          <SectionList
-            sections={data}
-            keyExtractor={(item, i) => item + i}
-            renderItem={({ item }) => <RadioButton name={item} />}
-            renderSectionHeader={({ section: { title } }) => (
-              <Text style={styles.title}>{title}</Text>
-            )}
-          />
-        </ThemedView>
-      </Modal>
-    </>
+    </View>
   );
 }
 
@@ -104,11 +51,6 @@ const styles = StyleSheet.create({
   },
   switch: {
     transform: [{ scale: 0.8 }],
-  },
-  title: {
-    textAlign: "center",
-    fontWeight: 500,
-    fontSize: sizes.fmd,
   },
 });
